se movieron css y test a componente extraido
